+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 
 -   Adds for For You tab with styled headers - luc
 -   Creates tab view container for new home - luc
+-   Adds Live and timed auctions in home view - luc
 
 ###### Consignments